This is an experimental fork of tangle.lisp to address two issues
1) The file that tangle generates is not independent of
literate-lisp. It still depends on defun-literate and with-web-syntax.

2) The code blocks have to be defined before they are used. However, a
commmon style (e.g. in TeXBook) is that code blocks are defined
*after* they are used

This modifies tangle-org-file to use text substitution instead of
relying on with-web-syntax. It does two passes over the org file. In
the first pass it reads all the code blocks. In the second pass it
writes the lisp file substituting the text for the code block
references.

Issues:

I've edited the lisp file, but not tangle.org. TBD. Ideally
have the asdf load the .org file instead of the .lisp file.

Tangling literate lisp is tricky - I protect defun-literate from being
replaced by defun - not sure if I have to do more.

The code blocks are still written out to the tangled file. They
probably shouldn't be as there is then a dependency on @= and @+=.

The non-code parts of the org file should probably become comments in
the lisp file.

In org-mode in emacs, when evaluating a function that references code
blocks, the code blocks will need to be scanned - probably each time either
evaluating a block, or once when org-babel-execute-buffer. So
some functions in emacs org mode need to be patched.