Make load of org files work, at least on ABCL, with code-block definitions allowed after they are referenced. It does this by modifying load to first read the file and gather the code blocks into named-code-blocks so they are available when the file is actually being loaded.
The second thing is that one doesn't need to use defun-literate, at least for
load (need to think about interactive eval). To do that we macrolet defun in the
body of the advised load. The macro calls expand-web-form before invoking the
real defun macro in a way I'm not sure is portable. Seems to work in ABCL and
SBCL.  SBCL locks the cl package, so we disable that lock for the whole
load. Would be better to just allow cl:defun to be rebound but I don't know how
to do that yet.

Third thing is that instead of @= and @+= I define :@= and :@+=. They can still
be defined as macros, but we don't have to worry about package this way.

With this :@= and :@+= are no-ops. They expand to nil. They are identified
without evaluation by the function org-file-gather-code-blocks.

--
Add a forgotten dependency on cl-ppcre to the system definition.