Bugfix:  named-code-blocks were being created, in the case of load and compile, by reading the found strings.  However, this reading was at the wrong time, and so there could be a package mismatch.  Instead save as strings and defer reading to  with-code-block in expand-web-form
Broke out tests.org into t/

Learned just enough ros to emulate one of the CI tests that jingtaozf had going (thanks!)
which is what found it.

Removed remnants of defun-literal and with-web-syntax

Added t/ros-test.sh with the test that found it. TBD: Expand.

Submitted https://github.com/jingtaozf/literate-lisp/issues/4 problem with
expand-web-form where form has a dotted list somewhere.

Protected compile-defun by checking that we are compiling or loading
and that the file is an ORG file. Not necessary, but no use doing
work we don't need to + protected against dotted list problem.